consumer/requirements.txt
kafka-python==2.0.2
snowflake-connector-python==3.7.0
